+++ packages/components/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.26.31](https://github.com/ttoss/ttoss/compare/@ttoss/components@1.26.30...@ttoss/components@1.26.31) (2023-04-13)
+
+**Note:** Version bump only for package @ttoss/components
+
 ## [1.26.30](https://github.com/ttoss/ttoss/compare/@ttoss/components@1.26.29...@ttoss/components@1.26.30) (2023-04-13)
 
 **Note:** Version bump only for package @ttoss/components