+++ packages/theme/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.4.14...@ttoss/theme@1.4.15) (2023-04-13)
+
+**Note:** Version bump only for package @ttoss/theme
+
 ## [1.4.14](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.4.13...@ttoss/theme@1.4.14) (2023-04-13)
 
 **Note:** Version bump only for package @ttoss/theme