+++ terezinha-farm/app/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.20](https://github.com/ttoss/ttoss/compare/@terezinha-farm/app@0.2.19...@terezinha-farm/app@0.2.20) (2023-04-13)
+
+**Note:** Version bump only for package @terezinha-farm/app
+
 ## [0.2.19](https://github.com/ttoss/ttoss/compare/@terezinha-farm/app@0.2.18...@terezinha-farm/app@0.2.19) (2023-04-13)
 
 **Note:** Version bump only for package @terezinha-farm/app